add docstrings and update uml

+++ VehicleCode/Input/DataCollector.h
@@ -1,3 +1,17 @@
+/*!
+ * @brief Read and collect data from sensors to send to the dashboard and car
+ * 
+ * Whenever checkSensors gets called, this class reads data from every sensor 
+ * that is ready to send data. It then creates a SensorData object for each 
+ * piece of data and sends it to the dashboard and car over CAN.
+ * 
+ * @param sensors (Sensor**) An array of pointers to Sensor objects
+ * @param numSensors (int) The number of sensors in the array
+ * @param startTime (unsigned long) The time that the program started in milliseconds
+ * 
+ * @return None 
+ */
+
 #ifndef DATACOLLECTOR_H
 #define DATACOLLECTOR_H
 
@@ -8,10 +22,10 @@ class DataCollector {
 private:
     // sensor objects, starting time, and number of sensors
     Sensor** sensors;
-    int timeZero;
+    unsigned long timeZero;
     int numSensors;
     
-    // CHECK new function
+    // CHECK new function for single use principle
     // Read signals from sensors
     void readData(Sensor* sensor);
